requirements.txt
# Core
langchain
langchain-openai
langchain-community
python-dotenv
openai

# API
fastapi
uvicorn[standard]
gunicorn
pydantic>=2.0

# Upstash Redis
upstash-redis

# Testing
pytest
pytest-asyncio
httpx